Add main.py shortdata.out
Added shortdata.out as a data source as a sample to analyze trade data. It is arranged in bytes and contains 8 doubles and 1 long int and it is arranged in big endian.

Added main.py to read in the data, reformat it into a 2-d matrix, check if there are any nan data, and clip maximum values down to 4. Timestamps that are generated by the dataset are compared to check if there is a consistent difference between each timestamp.