Update main.py
Added plot_histogram, but it does not work, I do not recommend calling that function.

Added SVM 2 regressors, one with a linear kernel and one with an RBF kernel. The 5th feature of data was chosen for training and testing. The data was split in an 80/20 ration and three values 0.001, 0.01, and 0.1 are chosen as the free parameter for training 6 models in total. A prediction is made on the dataset and the results are plotted against the test y for evaluation of the performance of the model.

The plot of the linear prediction is initially commented out because it is poor. To see the poor performance, uncomment the line.